+++ modules/browsers.nix
@@ -236,37 +236,37 @@
       blocked_install_message = "Install extensions via nix config!";
     };
     "uBlock0@raymondhill.net" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/latest/ublock-origin/latest.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4598854/ublock_origin-1.67.0.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;
     };
     "jid1-MnnxcxisBPnSXQ@jetpack" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/latest/privacy-badger17/latest.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4570378/privacy_badger17-2025.9.2.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;
     };
     "sponsorBlocker@ajay.app" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/latest/sponsorblock/latest.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4608179/sponsorblock-6.1.0.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;
     };
     "{1018e4d6-728f-4b20-ad56-37578a4de76b}" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/file/4567995/flagfox-6.1.90.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4609492/flagfox-6.1.92.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;
     };
     "jid1-KKzOGWgsW3Ao4Q@jetpack" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/latest/i-dont-care-about-cookies/latest.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4202634/i_dont_care_about_cookies-3.5.0.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;
     };
     "wayback_machine@mozilla.org" = {
-      install_url = "https://addons.mozilla.org/firefox/downloads/latest/wayback-machine_new/latest.xpi";
+      install_url = "https://addons.mozilla.org/firefox/downloads/file/4047136/wayback_machine_new-3.2.xpi";
       installation_mode = "force_installed";
       private_browsing = true;
       updates_disabled = true;